frontend: refactor index.js to have its functions placed in state folder
